#4281 lint + moving modals to admin tools
