+++ src/frontend/src/pages/AdminToolsPage/ProjectsConfig/RulesetTypeDeleteModal.tsx
